Added helper to list all blobs
