+++ src/BuildScriptGenerator/DotNetCore/VersionProviders/DotNetCoreSdkStorageVersionProvider.cs
@@ -57,13 +57,7 @@ namespace Microsoft.Oryx.BuildScriptGenerator.DotNetCore
             {
                 var httpClient = this.HttpClientFactory.CreateClient("general");
                 var sdkStorageBaseUrl = this.GetPlatformBinariesStorageBaseUrl();
-                var url = string.Format(SdkStorageConstants.ContainerMetadataUrlFormat, sdkStorageBaseUrl, DotNetCoreConstants.PlatformName, string.Empty);
-                var blobList = httpClient
-                    .GetStringAsync(url)
-                    .Result;
-
-                var xdoc = XDocument.Parse(blobList);
-                var marker = xdoc.Root.Element("NextMarker").Value;
+                var xdoc = ListBlobsHelper.GetAllBlobs(sdkStorageBaseUrl, DotNetCoreConstants.PlatformName, httpClient);
 
                 // keys represent runtime version, values represent sdk version
                 var supportedVersions = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase);
@@ -76,18 +70,6 @@ namespace Microsoft.Oryx.BuildScriptGenerator.DotNetCore
                     sdkVersionMetadataName = SdkStorageConstants.LegacySdkVersionMetadataName;
                     runtimeVersionMetadataName = SdkStorageConstants.LegacyDotnetRuntimeVersionMetadataName;
                 }
-
-                // if <NextMarker> element's value is not empty, we iterate through every page by appending marker value to the url
-                // and consolidate blobs from all the pages.
-                do
-                {
-                    url = string.Format(SdkStorageConstants.ContainerMetadataUrlFormat, sdkStorageBaseUrl, DotNetCoreConstants.PlatformName, marker);
-                    var blobListFromNextMarker = httpClient.GetStringAsync(url).Result;
-                    var xdocFromNextMarker = XDocument.Parse(blobListFromNextMarker);
-                    marker = xdocFromNextMarker.Root.Element("NextMarker").Value;
-                    xdoc.Descendants("Blobs").LastOrDefault().AddAfterSelf(xdocFromNextMarker.Descendants("Blobs"));
-                }
-                while (!string.IsNullOrEmpty(marker));
 
                 foreach (var metadataElement in xdoc.XPathSelectElements($"//Blobs/Blob/Metadata"))
                 {
